view 系のファイルを view パッケージに移動
