+++ app/src/main/res/layout/activity_game.xml
@@ -58,7 +58,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:orientation="vertical"/>
-        <xxx.prd.lastone.PaintView
+        <xxx.prd.lastone.view.PaintView
             android:id="@+id/paint_view"
             android:layout_width="match_parent"
             android:layout_height="match_parent" />
